plugins/verifica-cedolini/requirements.txt
pdfplumber>=0.10.0
pandas>=2.0.0
pyyaml>=6.0
